move tests to dfAggSuite
